change some s16 related to rupee counts into u16
- return type of `dComIfGs_getRupee` is now `u16` instead of `s16`
- type of field `mItemNowRupee` is now `u16` instead of `s16`

+++ include/d/d_save.h
@@ -15,7 +15,7 @@ public:
     void setSelectEquip(int i_no, u8 i_itemNo) { mSelectEquip[i_no] = i_itemNo; }
     u8 getWalletSize() { return mWalletSize; }
     void setWalletSize(u8 size) { mWalletSize = size; }
-    int getRupee() { return mRupee; }
+    u16 getRupee() { return mRupee; }
     void setRupee(u16 rupee) { mRupee = rupee; }
     u16 getRupeeMax() { return 0; } // TODO
     u8 getMagic() { return mMagic; }
